update for full separation
