+++ src/core/MRApp.js
@@ -15,7 +15,7 @@ import { MRSystem } from 'mrjs/core/MRSystem';
 import { AnimationSystem } from 'mrjs/core/componentSystems/AnimationSystem';
 import { ClippingSystem } from 'mrjs/core/componentSystems/ClippingSystem';
 import { ControlSystem } from 'mrjs/core/componentSystems/ControlSystem';
-import { CSSSystem } from 'mrjs/core/componentSystems/CSSSystem';
+import { StyleSystem } from 'mrjs/core/componentSystems/StyleSystem';
 import { LayoutSystem } from 'mrjs/core/componentSystems/LayoutSystem';
 import { MaskingSystem } from 'mrjs/core/componentSystems/MaskingSystem';
 import { PhysicsSystem } from 'mrjs/core/componentSystems/PhysicsSystem';
@@ -98,7 +98,8 @@ export class MRApp extends MRElement {
 
         this.layoutSystem = new LayoutSystem();
         this.textSystem = new TextSystem();
-        this.cssSystem = new CSSSystem();
+        this.geometrySystem = new GeometrySystem();
+        this.styleSystem = new StyleSystem();
         this.audioSystem = new AudioSystem();
 
         // initialize built in Systems
